refactor(FE): API 함수 경로 변경점 적용
