+++ front/src/widgets/ReviewCard.tsx
@@ -9,7 +9,7 @@ import {
 import { Link } from "react-router-dom";
 import { API_URL } from "@/shared/constants";
 import { useQuery } from "@tanstack/react-query";
-import { fetchUserInfoById } from "@/api/interaction";
+import { fetchUserInfoById } from "@/api/user";
 import { Separator } from "@/shared/shadcn-ui/separator";
 import { fetchReviewById } from "@/api/review";
 import { useSelector } from "react-redux";
